feat: 빅장 점수 1.5로 수정
Score: int -> double 수정

정수 or 1.5밖에 안들어오기 때문에 값이 이상하게 표현되지 않는다.

+++ src/main/java/janggi/domain/piece/Score.java
@@ -1,7 +1,7 @@
 package janggi.domain.piece;
 
 public record Score(
-        int value
+        double value
 ) {
     public static final Score NONE = new Score(0);
